Split the main script in three to have different cronjobs config
